requirements.txt
-e .
aws_cdk.aws_lambda==1.122.0
requests
boto3